Doing some reorganization of the code
