+++ build.gradle
@@ -17,4 +17,9 @@ dependencies {
     compile 'com.google.guava:guava:28.1-jre'
     compile group: 'org.mapdb', name: 'mapdb', version: '3.0.7'
     compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.28'
+    compile group: 'org.lz4', name: 'lz4-java', version: '1.6.0'
+    compile group: 'com.google.dagger', name: 'dagger', version: '2.0'
+    annotationProcessor 'com.google.dagger:dagger-compiler:2.20'
+    compileOnly 'org.projectlombok:lombok:1.18.10'
+    annotationProcessor 'org.projectlombok:lombok:1.18.10'
 }
