+++ src/main/java/com/brainlesslabs/momo/common/MomoBase.java
@@ -1,6 +1,9 @@
 package com.brainlesslabs.momo.common;
 
 import com.brainlesslabs.momo.common.exceptions.MomoException;
+import com.brainlesslabs.momo.common.options.Options;
+import com.brainlesslabs.momo.common.options.ReadOptions;
+import com.brainlesslabs.momo.common.options.WriteOptions;
 
 public interface MomoBase {
     void open(final Options options, Object... params) throws MomoException;
