--- src/main/java/com/brainlesslabs/momo/common/ReadOptions.java
@@ -1,4 +1,0 @@
-package com.brainlesslabs.momo.common;
-
-public interface ReadOptions {
-}
