--- src/main/java/com/brainlesslabs/momo/common/WriteOptions.java
@@ -1,4 +1,0 @@
-package com.brainlesslabs.momo.common;
-
-public interface WriteOptions {
-}
