+++ src/main/java/com/brainlesslabs/momo/common/options/ReadOptions.java
@@ -1,0 +1,4 @@
+package com.brainlesslabs.momo.common.options;
+
+public interface ReadOptions {
+}
