+++ src/main/java/com/brainlesslabs/momo/examples/Main.java
@@ -1,0 +1,13 @@
+package com.brainlesslabs.momo.examples;
+
+import com.brainlesslabs.momo.common.options.Options;
+import com.brainlesslabs.momo.impl.linklistheadcell.options.LinkListOptions;
+
+public class Main {
+    public static void main(String... options) {
+    }
+
+    public static void singleLinkedList(String... options) {
+        final Options opt = new LinkListOptions();
+    }
+}
