+++ src/main/java/com/brainlesslabs/momo/impl/linklistheadcell/options/LinkListOptions.java
@@ -1,0 +1,10 @@
+package com.brainlesslabs.momo.impl.linklistheadcell.options;
+
+import com.brainlesslabs.momo.common.options.Options;
+import lombok.AllArgsConstructor;
+import lombok.ToString;
+
+@AllArgsConstructor
+@ToString(callSuper = true)
+public class LinkListOptions extends Options {
+}
